requirements.txt
pytest
flake8
playwright
pytest-xdist
allure-pytest
python-dotenv
requests
pydantic